% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Reads in data from all Excel Files}
\usage{
getData(
  dird = "C:/Dev/R_Projects/MATH4753CURR0011/Data/K25936_Downloads/Excel/"
)
}
\arguments{
\item{dird}{this is the directory structure as a string}
}
\value{
a named list of the files read
}
\description{
This function will read a collection of xl files into R's workplace
}
\details{
This will greatly reduce the amount of work needed to perform analyses since the reading of files can be time consuming.
}
\examples{
\dontrun{
v<-getData()
}
}
