% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5.R
\name{getSample}
\alias{getSample}
\title{Lab 5 func}
\usage{
getSample(num0s = 5, num1s = 5, sampleSize = 10, replace = TRUE)
}
\arguments{
\item{num0s}{0}

\item{num1s}{1}

\item{sampleSize}{n}

\item{replace}{r}
}
\description{
Lab 5 func
}
