% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylab5.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter = 100, n = 10, p = 0.5, mkplot = TRUE)
}
\arguments{
\item{iter}{number of experiments consisting of n Bernoulli trials}

\item{n}{number of Bernouilli trials in each experiment}

\item{p}{probability of obtaining a success on each trial}
}
\value{
a table containing the frequency distribution of a random subset of size
'iter' from the collection of all 'n' Bernouilli trials
}
\description{
mybin
}
\examples{
\dontrun{
numExp = 65
numTrials = 10
probability = 0.6

mybin(iter = numExp, n = numTrials, p = probability)
}
}
