% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylab4.R
\name{myScatterhist}
\alias{myScatterhist}
\title{myScatterhist}
\usage{
myScatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{takes the x values of a distribution}

\item{y}{takes the y values of a distribution}

\item{xlab}{generates a label on the xaxis}

\item{ylab}{generates a label on the yaxis}
}
\value{
a histogram plotting values of Y with respect to X
}
\description{
myScatterhist
}
