% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{pareto}
\usage{
pareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{data}

\item{mn}{name of barplor}

\item{...}{}
}
\value{
returns a barplot with data
}
\description{
pareto
}
